pkmn-dataset/requirements.txt
datasets==4.4.1
pandas==2.3.3
sqlalchemy==2.0.44
python-dotenv==1.0.1
psycopg2-binary==2.9.11
Pillow==12.0.0
uv==0.9.8
matplotlib==3.10.7